refactor(Toolbar): make Toolbar SSR ready
