+++ packages/main/src/internal/ssr.ts
@@ -1,0 +1,15 @@
+import { EffectCallback, useEffect, useState } from 'react';
+
+export function useServerSideEffect(effect: EffectCallback) {
+  useEffect(effect, []);
+}
+
+export function useCanRenderPortal() {
+  const [allowed, setAllowed] = useState(false);
+
+  useEffect(() => {
+    setAllowed(true);
+  }, []);
+
+  return allowed;
+}
