--- packages/main/src/internal/useServerSideEffect.ts
@@ -1,5 +1,0 @@
-import { EffectCallback, useEffect } from 'react';
-
-export function useServerSideEffect(effect: EffectCallback) {
-  useEffect(effect, []);
-}
