+++ packages/main/src/internal/withWebComponent.tsx
@@ -13,7 +13,7 @@ import React, {
 } from 'react';
 import { CommonProps } from '../interfaces/CommonProps';
 import { Ui5DomRef } from '../interfaces/Ui5DomRef';
-import { useServerSideEffect } from './useServerSideEffect';
+import { useServerSideEffect } from './ssr';
 import { camelToKebabCase, capitalizeFirstLetter, kebabToCamelCase } from './utils';
 
 const createEventPropName = (eventName) => `on${capitalizeFirstLetter(kebabToCamelCase(eventName))}`;
